+++ NEWS.md
@@ -1,5 +1,19 @@
 # Release notes
 
+Version 0.4.1 (2024-03-15)
+--------------------------
+### Adjustment
+* The `print table` functionality is now available through the `export` button with the `REPL` option
+
+### Bugfix
+* Fixed issue with pinned and visible plots not being properly overwritten
+* Fixed issue with setting `expandAll = true` in the input argument for `GUI`
+* Fix issuewith non-existing export folder
+
+### Feature
+* Added functionality to export all JuMP variables to an excel file.
+* Added functionality to export all JuMP variables to the REPL using PrettyTables
+
 Version 0.4.0 (2024-03-12)
 --------------------------
 ### Adjustment